+++ src/main/java/com/project/eat/shop/ShopController.java
@@ -1,10 +1,7 @@
 package com.project.eat.shop;
 
 import lombok.extern.slf4j.Slf4j;
-import org.apache.tomcat.util.json.JSONParser;
-import org.apache.tomcat.util.json.ParseException;
-import org.json.JSONArray;
-import org.json.JSONException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -12,14 +9,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.URL;
-import java.nio.charset.StandardCharsets;
+
 import java.util.ArrayList;
 import java.util.List;
-import org.json.JSONObject;
+
 
 @Controller
 @Slf4j
@@ -495,7 +488,7 @@ public class ShopController {
 
     //shopDatil/num=${vo.shopId})
     @GetMapping("/shopDetail")
-    public String shopDetail(@RequestParam("num") int shopId, Model model){
+    public String shopDetail(@RequestParam("num") Long shopId, Model model){
         log.info("shopDetail called@#$@#$%#$@#!@#~~~~~~~~~~~~~~~~~~~~~~~~~");
         log.info("shopId: {}", shopId);
         model.addAttribute("shopId", shopId);
